Add CleanWorkflowProvider
