+++ client/src/commands/setup.ts
@@ -2,9 +2,8 @@ import { ExtensionContext } from "vscode";
 import { CommonLanguageClient } from "vscode-languageclient";
 import { CompareCleanWithWorkflowsCommand } from "./compareCleanWith";
 import { SelectForCleanCompareCommand } from "./selectForCleanCompare";
-import { CompareCleanWorkflowsCommand } from "./compareCleanWorkflows";
 import { PreviewCleanWorkflowCommand } from "./previewCleanWorkflow";
-import { CleanWorkflowDocumentProvider } from "../providers/cleanWorkflowDocumentProvider";
+import { CleanWorkflowProvider } from "../providers/cleanWorkflowProvider";
 
 /**
  * Registers all custom commands declared in package.json
@@ -13,14 +12,9 @@ import { CleanWorkflowDocumentProvider } from "../providers/cleanWorkflowDocumen
  */
 export function setupCommands(context: ExtensionContext, client: CommonLanguageClient) {
   context.subscriptions.push(new PreviewCleanWorkflowCommand(client).register());
-  context.subscriptions.push(new CompareCleanWorkflowsCommand(client).register());
   const selectForCompareProvider = new SelectForCleanCompareCommand(client);
   context.subscriptions.push(selectForCompareProvider.register());
   context.subscriptions.push(
-    new CompareCleanWithWorkflowsCommand(
-      client,
-      selectForCompareProvider,
-      new CleanWorkflowDocumentProvider(client)
-    ).register()
+    new CompareCleanWithWorkflowsCommand(client, selectForCompareProvider, new CleanWorkflowProvider(client)).register()
   );
 }
